chore: split client into multiple go files
